requirement.txt
django
python-decouple
djangorestframework
requests
djangorestframework-simplejwt
decouple